provisioning/requirements-tftp.txt
fbtftp
celery
redis
scrapli[ssh2]
scrapli[genie]
genie
jinja2
ruamel.yaml

pytest
black
flake8
bpython